requirements.txt
PyMuPDF==1.25.5
PyYAML==6.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
difflib
scikit-learn
gensim
transformers
torch
beautifulsoup4
fitz